* Uploading some of the scripts I use to test stuff. More to come soon.
I use WinDbg as my primary test 'target' as it is available natively for both x86 and x64, it's fast to open/close, and is available on all my dev machines. (Not to mention it's free and very easy to install if you don't already have it)

Just copy the entire 'Scripts' folder into your binary folder, and then drag and drop a script onto the appropriate batch file.